requirements.txt
torch
torchvision
matplotlib
numpy
streamlit
Pillow
PyYAML
transformers
datasets
